Puedo crear cualquier tipo de modelo, los ejemplos para PostMan estan el los DTO

+++ backend/src/main/java/com/mpl/backend/repository/EvaluacionItemRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mpl.backend.model.EvaluacionItem;
 
-public interface EvaluacionItemRepository extends JpaRepository<EvaluacionItem, Integer> {
+public interface EvaluacionItemRepository extends JpaRepository<EvaluacionItem, Long> {
 }
